Fixed CSS not loading for /dodge, close #11
The /dodge page was not loading CSS correctly, which was due to the route
name being the same a folder in the client/ folder which was specified as
the location for static content. So express was looking for
client/dodge/stylesheet.css, which doesn't exist.

A prefix was added to the static specification, so requests for static
content are now prefixed with static/ to avoid this issue.