Updated mixin calls and removed declarations
So there was a slightly new syntax for calling the mixins which i needed
to change. Then I removed mixin declarations from the script mixins, they
are now added using the include keyword which just copies the text from
the specified file in. As they are just a collection of scripts they don't
need to have a function/paramater to invoke, just c+p is fine.

Corrected the indentation of the scripts block in single.pug as it was
causing duplicate script includes.