Started attempting to implement BEM #43
So I needed some sort of guidelines to implement CSS a bit better. I
don't really know how to handle it well. I found a pattern called BEM
which I've had a crack at beginning to implement.

Added a doc with my notes on the general plan, and created a few new files
to handle moving the header block to the BEM style.