Implement BEM for lobby-list on /mingle #39
I know multiplayer isn't quite on the cards yet, but thought I could start
working on the lobby system I have, at least getting the css in a
reasonable place while I'm doing the rest of the site.

This meant copying the lobby system into mingle.pug and making a change to
the controller, to access the lobbyManager. Looks like it all works out of
the box which is pretty amazing.

So I've just broken down the first big block on this page, the lobby-list.
Working through the create and open blocks next.