The coverage edition
Got a bit carried away and haven't broken this down as I should've. But
basically spent the evening working on filling in the coverage holes. The
biggest gap was the server, and events module, which I hadn't started
looking at test-wise anyway.

So first filled some gaps in the client single player game, and then
worked on events. There was a lot of dicking around with sockets, but got
them working in the end, and have some coverage now happening, and can see
how most of it can get done.