Resolve #47
I was getting a warning abouve having too many connection listeners added.
I figured this was becuase I was trying to tear down and build the server
before and after every test, but the process isn't completely shut down
with the way I was doing that. So the server would start again, and
register a namespace listener in the events, but the old one was still
there so a few tests was enough to hit the limit.

I've now just set it up so the server gets created at the top of the spec,
and stays on for all tests.

Until I can find a way to properly tear it down and remove all listeners
before restarting, if that's even a reasonable, worthwhile thing to do.