Resolve #48
I was having issues with sockets being added to a room multiple times
somehow, and I think in the end my problem was socket connection issues.

I thought I'd set the client_rcv to call join on a lobby, only once the
newLobby event was rcvd. But this meant setting client_rcv.on('newLobby'
etc, which meant all the tests for join were listening to all the newLobby
events, some of which are in the open test group.

Not really 100% sure how things were getting so overconnected, but it's
working now.