Resolved tests and updated middleware
Probably should've been two commits.

First I was getting errors running the server because errorhandler and
logger used to be included middleware with express, and now they're
separate modules.

So express.errorHandler was replaced with the errorhandler package.

express.logger was replaced with morgan.

Then there were test failures due to listen being called on the server
when it was already active. To fix this I put server spinup/teardown in
berforeAll and afterAll functions.

These are available in jasmine >2.0 and I had an older package
jasmine-node installed. So had to replace that package and this works
now.