Added documentation, coverage, and fixed browserify
Updated the readme to describe how working on this project should look
to hopefully make it easier to get into if I have another break, or for
other people!

Amended istanbul to nyc, as istanbul is deprecated.

gitignored the nyc logging folder.

Added a browserify file to list requires, rather than listing them all
in the script. watchify uses this file too.